+++ CHANGELOG.md
@@ -12,8 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `ITest` to runs before `Release`
 - Update `IGitRelease` with custom method to create release and pre-release.
 - Update `GitHubExtension` to release or pre-release.
+- Update order of `Targets`
 ### Added
 - Add `IGitPreRelease` with custom method to create pre-release when version is pre-release.
+- Add `IPrePack` with custom method to create pre-release when version is pre-release.
 
 ## [1.6.1] / 2023-10-02
 ### Fixed